refactor: Be more specific about datatypes, also move remaining stat parsing into playermanager
This change is to reduce casting and parsing stats and info from strings, as the parser already pulls out the values into the correct datatype

+++ CrossfireCore/NewClient.cs
@@ -164,7 +164,7 @@ namespace CrossfireCore
 
 
         [Flags]
-        public enum ItemFlags : uint
+        public enum ItemFlags : UInt32
         {
             None = 0,
 
@@ -219,7 +219,7 @@ namespace CrossfireCore
         }
 
         [Flags]
-        public enum StatFlags : uint
+        public enum StatFlags : UInt16
         {
             Firing = 0x01,
             Running = 0x02,
